we add edit to product functionality

+++ app/action/delete.ts
@@ -1,17 +1,19 @@
-// app/action/delete.ts
-'use server'
+// app/actions/delete.ts
+'use server';
+
 import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 
-export async function deleteProduct (id:any):Promise<void> {
-    try {
-        await prisma.products.delete({
-            where: {id}
-        })
-    revalidatePath("/products")
-    }
-    catch(err) {
-        console.log("The error is",err)
-        throw err
-    }
+export async function deleteProduct(formData: FormData) {
+  const id = formData.get("id")?.toString();
+
+  if (!id || id.trim() === "") {
+    throw new Error("Invalid ID");
+  }
+
+  await prisma.products.delete({
+    where: { id },
+  });
+
+  revalidatePath("/products");
 }
